job-analysis-tool/backend/requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
python-multipart
pymysql
requests